+++ bagel/codeanalysis/binding/boundbinaryoperatorkind.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class BoundBinaryOperatorKind(Enum):
+    ADDITION = 0
+    SUBTRACTION = 1
+    MULTIPLICATION = 2
+    DIVISION = 3
