+++ bagel/codeanalysis/binding/boundexpression.py
@@ -1,0 +1,9 @@
+from .boundnode import BoundNode
+
+
+class BoundExpression(BoundNode):
+    _type: object
+
+    @property
+    def type(self):
+        return self._type
