+++ bagel/codeanalysis/binding/boundnode.py
@@ -1,0 +1,9 @@
+from .boundnodekind import BoundNodeKind
+
+
+class BoundNode:
+    _kind: BoundNodeKind
+
+    @property
+    def kind(self) -> BoundNodeKind:
+        return self._kind
