+++ bagel/codeanalysis/binding/boundnodekind.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class BoundNodeKind(Enum):
+    LITERALEXPRESSION = 0
+    UNARYEXPRESSION = 1
+    BINARYEXPRESSION = 2
