+++ bagel/codeanalysis/binding/boundunaryoperatorkind.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class BoundUnaryOperatorKind(Enum):
+    IDENTITY = 0
+    NEGATION = 1
